Updated FriendCapacity to use u128 instead of u64.

+++ cswitch/src/inner_messages.rs
@@ -434,8 +434,8 @@ enum NetworkerToFunder {
 // ------------------------
 
 pub struct FriendCapacity {
-    send: u64,
-    recv: u64,
+    send: u128,
+    recv: u128,
 }
 
 #[derive(Debug, Eq, PartialEq, Clone)]
